% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createAB.old}
\alias{createAB.old}
\title{Creates a "Arbeitsblatt" with and without solution from a Rmd-file}
\usage{
createAB.old(infile)
}
\arguments{
\item{infile}{the rmarkdown description file}
}
\description{
Creates a "Arbeitsblatt" with and without solution from a Rmd-file
}

