% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeRMD_tolk.R
\name{mergeRMD2}
\alias{mergeRMD2}
\title{Merges Rmd-files into a single 'Arbeitsblatt'}
\usage{
mergeRMD2(
  mergedFileName = "book.Rmd",
  title = ".",
  files,
  preamble,
  print.paths = FALSE,
  print.newpage = TRUE,
  print.names = TRUE
)
}
\arguments{
\item{mergedFileName}{Name of the file containing all merged files, along with its filepath. When filepath is not given, 
current working directory is assumed.}

\item{title}{Title of the merged collection of files (e.g. Arbeitsblattname).}

\item{files}{List of files containing the Rmd files to be merged. Should be provided as c("file path1","file path2", ...). 
See examples on interactive file selection in the details.}

\item{preamble}{(Optional) File path to the preamble written in plain Rmd file (meta-text, R code allowed). Preamble will be placed before all exercises. Latex packages should be placed in header_lsg.tex.}

\item{print.paths}{(By default, FALSE) Indicate whether to print paths of each exercise (provided as clickable links in the final pdf document).}

\item{print.newpage}{(By default, TRUE) Indicate whether to start each exercise from a new sheet.}
}
\description{
Merges Rmd-files into a single 'Arbeitsblatt'
}
\details{
Merges Rmd files from different locations into one Rmd file. Relies on packages rmarkdown, knitr, devtools, data.table, idpedu (GitHub).

It is important to state **options(useFancyQuotes = FALSE)** before running the function. 
It is also preferred to leave free the last line in each .Rmd file to avoid problems with **readLines()** command, which is used in mergeRMD2.
The data and figures called in each individual Rmd file must be placed in the same directory as the individual Rmd file.
Each exercises with header "Aufgabe" is automatically numbered in the same order as in the **files** argument.
}
\examples{
\dontrun{

mergeRMD2(mergedFileName = "V:/tolk/Private/PROJECT_02_DUERR/TEST_FILES_2/DUMMY/book4.Rmd",
         title = "Arbeitsblatt 1",
         files = c("V:/tolk/Private/PROJECT_02_DUERR/TEST_FILES_2/DUMMY/HA01.Rmd", 
                   "V:/tolk/Private/PROJECT_02_DUERR/TEST_FILES_2/DUMMY/HA02.Rmd"),
         preamble="V:/tolk/Private/PROJECT_02_DUERR/TEST_FILES_2/DUMMY/Preamble.Rmd" ,
         print.paths=T, 
         print.newpage=T)

createAB("V:/tolk/Private/PROJECT_02_DUERR/TEST_FILES_2/DUMMY2/book4.Rmd") 


#======================================================
# Ways to select files and directories
         
# 1. Interactively choose directory with GUI window, then apply files crawler
path1=gsub("\\\\\\\\", "/", choose.dir())
         
## or choose specific files
files1=gsub("\\\\\\\\", "/", choose.files())
         
# 2. Replace the address line below with the raw path provided by Windows.
         
# 3. Copy the text of the file(directory) path from the desired location and run the following
path1=gsub("\\\\\\\\", "/", readClipboard())
         
# Files Crawler
 (files1=list.files(path =path1, pattern = "*.Rmd", full.names=TRUE, recursive = TRUE))

#======================================================
                    
         
 }
}
\author{
Vasily Tolkachev \email{tolk@zhaw.ch}, Oliver Durr \email{oliver.duerr@zhaw.ch}
}
