% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MergeRMD.R
\name{mergeRMDFiles}
\alias{mergeRMDFiles}
\title{Takes several rmd-files and merge it into a single file}
\usage{
mergeRMDFiles(dir = ".", title = ".", files, mergedFileName = "book.Rmd")
}
\arguments{
\item{dir}{a temporary working directort}

\item{title}{of the exercise sheet}

\item{files}{a list of files containing the rmd-files which will be merged

To create a list of file one might use
 files = list.files(path='V:/tolk/Private/PROJECT 02 O. DUERR/TEST FILES',pattern = "HA0[1-2].Rmd")}
}
\description{
Takes several rmd-files and merge it into a single file
}

