% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createAB}
\alias{createAB}
\title{Creates 2 excersis sheets with and without solution from a r-makeuo file
All the compiling happens in a temporary folder (also better then using dropbox or the like)}
\usage{
createAB(infile, wast1 = FALSE)
}
\arguments{
\item{infile}{the rmarkdown description file}
}
\description{
Creates 2 excersis sheets with and without solution from a r-makeuo file
All the compiling happens in a temporary folder (also better then using dropbox or the like)
}

