% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createAnkreuzBlatt}
\alias{createAnkreuzBlatt}
\title{Creates an "Ankreuzblatt" from a file containing the student names}
\usage{
createAnkreuzBlatt(infile, output_file, header, name)
}
\arguments{
\item{infile}{tab-delimited file containig the name of the stundent col 1 and how often he showed his excercise}

\item{outfile}{the name of the output file}

\item{header}{the column names of the sheet e.g. header = c("Aufgabe 1 a", "Aufgabe 1 b")}

\item{name}{the name of the sheet}
}
\description{
Creates an "Ankreuzblatt" from a file containing the student names
}

