% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PraktikumsUtils.R
\name{randomPick}
\alias{randomPick}
\title{Picks a student by random from a file containing the student names and how often they have shown their solution}
\usage{
randomPick(infile)
}
\arguments{
\item{infile}{the file containing the names and how often they have shown their solution}
}
\description{
Picks a student by random from a file containing the student names and how often they have shown their solution
}

